+++ CHANGELOG.md
@@ -27,7 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New arena building in desert cities, suitable for PVP, also NPCs like to watch the fights too
 - The loading screen now displays status updates for singleplayer server and client initialization progress
 - New Frost Gigas attacks & AI
-- Weapons and armor load from plugins
+- Allow plugins to add weapon and armor items
 
 ### Changed